Move exceptions & commands to `automation` package
In the .NET SDK, it was likely a mistake having exceptions in a separate `Exceptions` namespace. With commands, it doesn't buy us that much having it in a separate package. This does add some additional types in the main package, but there's already a bunch of stuff there, so it's not making things much worse.
