+++ sdk/java/pulumi/src/main/java/com/pulumi/experimental/automation/AutomationException.java
@@ -1,6 +1,6 @@
 // Copyright 2025, Pulumi Corporation
 
-package com.pulumi.experimental.automation.exceptions;
+package com.pulumi.experimental.automation;
 
 public class AutomationException extends Exception {
     public AutomationException(String message) {
