+++ sdk/java/pulumi/src/main/java/com/pulumi/experimental/automation/CommandException.java
@@ -1,0 +1,9 @@
+// Copyright 2025, Pulumi Corporation
+
+package com.pulumi.experimental.automation;
+
+public class CommandException extends AutomationException {
+    public CommandException(CommandResult result) {
+        super(result.toString());
+    }
+}
