+++ sdk/java/pulumi/src/main/java/com/pulumi/experimental/automation/ProjectSettingsConflictException.java
@@ -1,6 +1,6 @@
 // Copyright 2025, Pulumi Corporation
 
-package com.pulumi.experimental.automation.exceptions;
+package com.pulumi.experimental.automation;
 
 /**
  * Thrown when creating a Workspace detects a conflict between
