+++ sdk/java/pulumi/src/main/java/com/pulumi/experimental/automation/StackNotFoundException.java
@@ -1,8 +1,6 @@
 // Copyright 2025, Pulumi Corporation
 
-package com.pulumi.experimental.automation.commands.exceptions;
-
-import com.pulumi.experimental.automation.commands.CommandResult;
+package com.pulumi.experimental.automation;
 
 public class StackNotFoundException extends CommandException {
     public StackNotFoundException(CommandResult result) {
